Restructured model package and fixed a circular dependency issue
caused by it
