+++ glotaran/__init__.py
@@ -1,7 +1,6 @@
 """Glotaran package __init__.py"""
 
 from . import io  # noqa: F401
-from . import model  # noqa: F401
 from . import parameter
 
 __version__ = "0.2.0"
