+++ glotaran/model/base_model.py
