+++ glotaran/model/property.py
