+++ glotaran/parse/register.py
@@ -1,7 +1,12 @@
 """A register for models"""
-import typing
+from __future__ import annotations
 
-from glotaran.model.model import Model
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from typing import List
+
+    from glotaran.model import Model
 
 _model_register = {}
 
@@ -41,5 +46,5 @@ def get_model(model_type: str) -> Model:
     return _model_register[model_type]
 
 
-def known_model_names() -> typing.List[str]:
+def known_model_names() -> List[str]:
     return [name for name in _model_register]
